- message.component.ts (line 18-22) Allows the component to detect whether a specific layout class has been applied to it, and store that information in the isLayout01 property.
This information could then be used by the component's template or other parts of the application to conditionally render or style the message based on the layout.

- message.component.html (line 1-3 & 9)
Conditionally display different SVG icons based on the value of the isLayout01 property. If isLayout01 is false, the first SVG icon will be displayed. If isLayout01 is true, the second SVG icon will be displayed.

- app.component.html (line 14 & 19)
Create an instance of the MessageComponent and apply a specific CSS class to it (layout--01 or layout--02).